+++ symmetria/__init__.py
@@ -1,6 +1,6 @@
 from symmetria.elements.cycle import Cycle
 from symmetria.elements.permutation import Permutation
-from symmetria.generators.random.api import random, random_generator
+from symmetria.generators.random.api import random, random_generator, random_permutation
 from symmetria.generators.algorithm.api import generate, permutation_generator
 from symmetria.elements.cycle_decomposition import CycleDecomposition
 
@@ -10,6 +10,7 @@ __all__ = [
     "permutation_generator",
     "random",
     "random_generator",
+    "random_permutation",
     "Permutation",
     "Cycle",
     "CycleDecomposition",