+++ symmetria/generators/random/api.py
@@ -1,3 +1,4 @@
+import warnings
 from typing import List, Generator
 
 from symmetria import Permutation
@@ -9,7 +10,7 @@ from symmetria.generators.random._algorithms import (
     _fisher_yates_shuffle_generator,
 )
 
-__all__ = ["random", "random_generator"]
+__all__ = ["random", "random_generator", "random_permutation"]
 
 _SUPPORTED_ALGORITHM: List[str] = [
     "random",
@@ -19,6 +20,8 @@ _SUPPORTED_ALGORITHM: List[str] = [
 
 def random(degree: int, algorithm: str = "random") -> Permutation:
     """Generate a random permutation of the given degree based on the chosen algorithm.
+
+    :warning: This function will be deprecated in a future version. Use 'random_permutation' instead.
 
     The method generate a random permutation of the given degree using the specified algorithm.
 
@@ -37,6 +40,37 @@ def random(degree: int, algorithm: str = "random") -> Permutation:
         >>> import symmetria
         ...
         >>> permutation = symmetria.random(degree=3, algorithm="fisher-yates")
+    """
+    warnings.warn(
+        "This function will be deprecated in a future version. Use 'random_permutation' instead.",
+        PendingDeprecationWarning,
+        stacklevel=1,
+    )
+    _check_algorithm_parameter(value=algorithm, supported=_SUPPORTED_ALGORITHM)
+    _check_degree_parameter(value=degree)
+    return _relevant_random_permutation(algorithm=algorithm, degree=degree)
+
+
+def random_permutation(degree: int, algorithm: str = "random") -> Permutation:
+    """Generate a random permutation of the given degree based on the chosen algorithm.
+
+    The method generate a random permutation of the given degree using the specified algorithm.
+
+    :param degree: The degree of the permutation to be generated. Must be a non-zero positive integer.
+    :type degree: int
+    :param algorithm: The algorithm to use for generating permutations.
+        Default is ``random``.
+    :type algorithm: str, optional
+
+    :return: A permutation object.
+    :rtype: Permutation
+
+    :raises ValueError: If the algorithm is not supported or the degree is invalid.
+
+    :examples:
+        >>> from symmetria import random_permutation
+        ...
+        >>> permutation = random_permutation(degree=3, algorithm="fisher-yates")
     """
     _check_algorithm_parameter(value=algorithm, supported=_SUPPORTED_ALGORITHM)
     _check_degree_parameter(value=degree)
@@ -68,10 +102,16 @@ def random_generator(degree: int, algorithm: str = "random") -> Generator[Permut
     :raises ValueError: If the algorithm is not supported or the degree is invalid.
 
     :examples:
-        >>> import symmetria
+        >>> from symmetria import random_generator
         ...
-        >>> permutations = symmetria.random_generator(degree=3, algorithm="fisher-yates")
+        >>> permutations = random_generator(degree=3, algorithm="fisher-yates")
     """
+    warnings.warn(
+        "The API of this method will be deprecated in a future version. \n"
+        "Use 'from symmetria import random_generator' instead of 'import symmetria.",
+        PendingDeprecationWarning,
+        stacklevel=1,
+    )
     _check_algorithm_parameter(value=algorithm, supported=_SUPPORTED_ALGORITHM)
     _check_degree_parameter(value=degree)
     return _relevant_random_generator(algorithm=algorithm, degree=degree)
@@ -83,12 +123,3 @@ def _relevant_random_generator(algorithm: str, degree: int) -> Generator[Permuta
         return _random_shuffle_generator(degree=degree)
     elif algorithm == "fisher-yates":
         return _fisher_yates_shuffle_generator(degree=degree)
-
-
-if __name__ == "__main__":
-    print(random(3, "fisher-yates").rep())
-    for idx, p in enumerate(random_generator(12, "fisher-yates")):
-        print(p.rep())
-
-        if idx == 10:
-            break